+++ src/package/themes/default/components/skeleton.ts
@@ -1,0 +1,9 @@
+export default {
+    width: '100%',
+    height: '1rem',
+    lineHeight: '0.9rem',
+    borderRadius: 'var(--vf-radii-md)',
+    backgroundColor: 'var(--vf-gutter-color)',
+    shimmerColor: 'var(--vf-bg-color)',
+    animationDuration: '2s',
+};