+++ src/package/themes/default/index.ts
@@ -18,6 +18,7 @@ import radio from './components/radio';
 import switchComponent from './components/switch';
 import tabs from './components/tabs';
 import tooltip from './components/tooltip';
+import skeleton from './components/skeleton';
 import toast from './components/toast';
 import alert from './components/alert';
 
@@ -44,6 +45,7 @@ export default {
         switch: switchComponent,
         tabs,
         tooltip,
+        skeleton,
         toast,
         alert,
     },